Improve cluster helper
The add_and_rebalance method of the ClusterOperationHelper
doesn't need to have the password as a parameter, as it
can get it from the master server object.

This simplifies the call and makes the whole method less
fragile. The username isn't hardcoded to "Administrator"
anymore.

The logging was improved, the port of the server is now
part of the output.

Change-Id: I482d18f0d836db7c24926aa8977392eb28ceb2eb
Reviewed-on: http://review.couchbase.org/12533
Reviewed-by: Keith Batten <kbatten@gmail.com>
Reviewed-by: Tommie McAfee <tommie@couchbase.com>
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: Volker Mische <volker.mische@gmail.com>