Make spatial tests work with Python 2.7
It isn't a good idea to subclass unittest.TestCase for
the SpatialHelper class, as it is not really a test case.
The better approach is to initiate the SpatialHelper class
with the original test case. This way we can still have
assertions.

Previously the tests based on SpatialHelper only worked
with Python 2.6. This change makes it compatible with
Python 2.7.

Change-Id: I04b0cc438968bd766424f38c5d51a83e4a3b1eff
Reviewed-on: http://review.couchbase.org/12621
Reviewed-by: Tommie McAfee <tommie@couchbase.com>
Tested-by: Volker Mische <volker.mische@gmail.com>
Reviewed-by: Volker Mische <volker.mische@gmail.com>