Add _verify_keys for faster verification
In my quick tests the _verify_keys function is about 10x faster
than the  _verify_docs_doc_name. It doesn't need to create a
MapReduce View.

This commit also changes _get_key to always return ASCII values.
This makes the comparison with the inserted keys easier (else
we need to convert the doc_names to unicode) and we have ASCII
compatible keys anyway.

Change-Id: Ia07254e4fba5385d64e19fa12ae984d07d7fde5f
Reviewed-on: http://review.couchbase.org/12950
Reviewed-by: Karan Kumar <karan@couchbase.com>
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: Volker Mische <volker.mische@gmail.com>