Deterministic, random-looking get/update key accesses for mcsoda.
This is the new default behavior of mcsoda (also, when random=0).

Aliaksey Kandratsenka described this formula to generate the next
command's key (for gets and updates) by updating a position...

  position = (position + LARGE_PRIME) % num_items

When num_items is stable, position should visit all the values in
range [0, num_items), as long as LARGE_PRIME and num_items are
relatively prime to each other.

Change-Id: I22440a03bd07c5cea781840b92c1fe9833748296
Reviewed-on: http://review.couchbase.org/13448
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Karan Kumar <karan@couchbase.com>