Added three checkpoint test cases
This commit adds test cases that test to see if a checkpoint is
properly created after inserting a certain number of items, to
see if a new checkpoint is created after a checkpoint timeout,
and to make sure that checkpoint deduplication is properly working.

This commit also includes a new stats helper file and the
configuration file for running checkpoint tests.

Change-Id: I4122356e668b0a22ad236c98b36b297b41bccc6f
Reviewed-on: http://review.couchbase.org/15077
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>