Refactored KVStore and document generation tasks
This is a rather large change and address a few areas that are
all pretty connected.

KVStore:
This change completly re-writes the kvstore to allow it to run
with multiple threads accessing it at the same time. This is
accomplished with a partition strategy which means that touching
the kvstore only requires locking part of it.

Data loding tasks:
The data loading taks here allow a user to run data loads over
certain key spaces or allow the user to randomly do operations
on key that are contained in the kvstore.

VBucketAwareMemcached:
I also added some functions here that will allow you to do a few
of the memcached operation with only a single call. Before you
would have to ask for a specific connection and then do the operation.
This addition also includes exponential backoff code and handles
not my vbucket errors.

Change-Id: Ide2b91f1961470766853450b67485700c9e87fc9
Reviewed-on: http://review.couchbase.org/15514
Reviewed-by: Karan Kumar <karan@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>