MB-100: cmd_get graph climbing-up issue
First of all, this might affect current eperf results, \
because mcsoda used num_ops as one of the inputs \
for key generation.

For clients other than master (prefix=0), \
mcsoda caculated and passed in offset (start_at) \
for all GET cmds. This led to a bump of \
doing all SETs no GETs in the beginning of access phase.

Change-Id: I52152cd841370432245b2a9c1318ade8028a4436
Reviewed-on: http://review.couchbase.org/17246
Reviewed-by: Pavel Paulau <pavel.paulau@gmail.com>
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Steve Yen <steve.yen@gmail.com>
Tested-by: Ronnie Sun <ronnie@couchbase.com>