Fix for wait_for_memcached
wait_for_memcached earlier threw an error,
when a bucket is created for a cluster after rebalancing.
The Fix: Removed redundant code from wait_for_memcached,
and added functionality in wait_for_vbuckets_ready_state()
that will cover both scenarios of when executed using
cluster_run or on a set of vms.
Replaced the call to MemcachedClientHelper with a call
to MemcachedClient itself.
Patch 18: Incorporated all the changes pointed out in
patch16 and patch17.

Change-Id: I4048c58cd8d1ee6a58845f2fda17b57929bc9981
Reviewed-on: http://review.couchbase.org/18228
Reviewed-by: Andrei Baranouski <andrei.baranouski@gmail.com>
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: Volker Mische <volker.mische@gmail.com>