Systest Platform
Initial check in of platform for system testing.
Adding pysystest directory which contains
a python-celery application that operates as a
worker for performing tasks.  Package also
consists of a cbsystest.py command line interface
for defining and running new workloads.
Currently python and ruby are supported sdks
for handling underlying requests to the couch cluster.

Environmental Vars:
  testcfg.py

Starting celery worker with scheduler(-B):
  celery worker -A app -B

Running a workload:
1) Define simple document template:
  python cbsystest.py import template --name "default" --kvpair ' "hello":"world"
2) Run workload using predifined template at 1k ops/sec
  python cbsystest.py run workload  --create 50 --get 50  --ops 1000 --cc_queues docs_to_consume --template "default

Change-Id: I5979b5f9ac3ddc4581f7f64899abdd4c8166a510
Reviewed-on: http://review.couchbase.org/19019
Reviewed-by: Tommie McAfee <tommie@couchbase.com>
Tested-by: Tommie McAfee <tommie@couchbase.com>