CBQE-379 Create new test cases for CAS functional test
CAS value manipulation by update, delete, expire test.
We need to verify whether the memchached client operations
set, get, and delete works properly and retun a valid CAS value.

For delete(), after it is called, the item cas is always reset to zero.
So we need to figure out a way to do verification about delete

Change-Id: I15b5de1e0e82e78f924905f694b99914b284ed7a
Reviewed-on: http://review.couchbase.org/19120
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Karan Kumar <karan@couchbase.com>