MB-100: add checking seqno=0 in revision verification
The seqno should never be 0 in anycase, today the verification code
jsut compares the seqno on source and destination side but does not
check if it is 0 or not. This change-set will fix that and also add
logs on the progress of verification, which may take quite long for
large number items.

Change-Id: Ib4028cc020a34664cdad84c015d760c9f4f55f05
Reviewed-on: http://review.couchbase.org/19581
Reviewed-by: Ketaki Gangal <ketakigangal@gmail.com>
Tested-by: Junyi Xie <junyi.couchbase@gmail.com>