CBQE-464 add cli test (cbcollect_info and cbepctl)
Add cbcollect_info and cbepctl from cli tools as a new automation job.
In this cbcollect, we load some items into cluster and use cbcollect_info
to generate a log zip file. We verify the whether all the logs generated
in the zip file and check whether all the bucket info is in stats.log
In cbepctl, we verify the result based on the output of the command line.
Also the test can only be on LINUX right now

Change-Id: I141f6bf72add8fbc0beba90ed2885d47b9589b2a
Reviewed-on: http://review.couchbase.org/19686
Reviewed-by: Chisheng Hong <chisheng@couchbase.com>
Tested-by: Chisheng Hong <chisheng@couchbase.com>