CBQE-1: allow recv method to terminate on '\0' ch
allow clients to send '\0' characters to notify
sdkserver that not more data is to be sent.

currently the sdkserver knows that clients have
finished sending data because the connection closes.
However, if the client remains open and wants
to receive a response from the sdkserver it can now
send '\0' then call recv.

Change-Id: Iad7ba2c3e5082ff07bdbbdb0d4e618b57d94bc21
Reviewed-on: http://review.couchbase.org/25112
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Chisheng Hong <chisheng@couchbase.com>