CBQE-1182: recover non-deleted items from deactivated workload
This fixes an issue for the case that you want to delete all created documents.
Problem is the workload_manager is very proactive about queueing up a bunch of tasks before they are scheduled to run.  This pulls doc_ids from the queue.  If the workload is stopped (say post condition met), then the pulled doc_ids are lost and can never be deleted.
Here we requeue the docs when workload is deactivated

Change-Id: Ia391e5d2440e343a09d9c0eebdb45834fdb4b9d6
Reviewed-on: http://review.couchbase.org/25404
Reviewed-by: Chisheng Hong <chisheng@couchbase.com>
Tested-by: Tommie McAfee <tommie@couchbase.com>