CBQE-1183 able to define number of initial nodes for clusters(include xdcr usage)
In system test automation, when we setup the initial cluster,
we sometimes won't use all the nodes defined in our ini file.
Like view ssd linux, for example, we will have 9 nodes in ini file
but we only want to start the test with 7 in cluster.

Cluster status is able to remember all the available nodes info
for the system test

And Due to this change http://review.couchbase.org/#/c/25405/,
user may restart and continue the system test with -I app.init
But we only want to setup the cluster just once, at the beginning
of the test. So we make cluster setup true or false option as celery
command option.

If you want to setup the cluster, celery worker command is:
celery worker -A app -B -c 8 -l ERROR --purge -I app.init,app.setup

Change-Id: Ib54705bd225ec2fb4ffd554c2abb17b561a969c6
Reviewed-on: http://review.couchbase.org/25422
Reviewed-by: Tommie McAfee <tommie@couchbase.com>
Tested-by: buildbot <build@couchbase.com>