MB-8295: Test dev views
This change will make it possible to test dev views.

Development views just return a subset of the data. When you
query a dev view with "full_set=true", it will use the
production view and return the full set.

Hence I disable setting "full_set" by default, so that we can
trigger dev views.

Production views will also be used if the number of documents
is to low. Therefore the test_dev_view() puts additional documents
into the database to overcome this amount.

Change-Id: Icbc49bb2a2043e87a29feb7b2ff5d17ff8d99664
Reviewed-on: http://review.couchbase.org/26520
Reviewed-by: Iryna Mironava <irynamironava@yandex.ru>
Reviewed-by: Filipe David Borba Manana <fdmanana@gmail.com>
Tested-by: buildbot <build@couchbase.com>