CBQE-2338: MemcachedError: Memcached error #7 'Not my vbucket' wrongly raised by testrunner code
async_load_all_buckets for some reason had default batch_size as 1 which caused Set() to be invoked instead of SetMulti(). Data helper's Set() is designed to throw an exception if NOT_MY_VBUCKET is encountered more than 5 times per vbucket. When the server is rebalancing,this is common. SetMulti() handles this correctly. Therefore changing batch size to 1000 to allow batch loading and correct handling of
this memcached error.

Change-Id: Icd0bbbc043b937372dcb22f23c217a87d5e7758c
Reviewed-on: http://review.couchbase.org/40066
Reviewed-by: Sangharsh Agarwal <sangharsh.agarwal@globallogic.com>
Tested-by: Sangharsh Agarwal <sangharsh.agarwal@globallogic.com>