Don't flood the console with not my vbucket msgs
During "wait_for_vbuckets_ready_state" we might receive a
lot of "not my vbucket responses" while we're waiting for
the cluster to make the vbuckets available on the nodes.
There is no point of flooding the console with a ton of such
messages (in my small test I received 80+ lines with the
following text).

2014-08-12 07:56:40 | ERROR | MainProcess | Cluster_Thread |
[bucket_helper.wait_for_vbuckets_ready_state] : Memcached error #7 'Not my
vbucket...

Change-Id: I151f2ce3b0c07367a84b41c1e5c24b0dd5cf772a
Reviewed-on: http://review.couchbase.org/40519
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Tommie McAfee <tommie@couchbase.com>