Retry bucket creation
When running the test suite on Windows it have sporadic test
failures due to the fact that sometimes the bucket deletion
takes more time than expected.

I believe that we need to improve the bucket deletion code
to correctly monitor when the bucket is completely gone so
that it may be recreated, since this workaround will fail
if the bucket deletion code is really slow.

Change-Id: I0b8b1385483fdccb525f7d88d3435273ec4755e1
Reviewed-on: http://review.couchbase.org/40810
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Volker Mische <volker.mische@gmail.com>