CBQE-2403:: Implement and add Checks for Bad Replicas in Rebalance and Failover tests
Bad Replicas condition can happen after we do rebalance and/or failover and this needs to be traced at the end of our tests.

Here is a bug which can be tracked with bad replica condition check

https://www.couchbase.com/issues/browse/MB-12100

We have added a sample usage test as well.

analysis.clusterinfoanalysis.DataAnalysisTests.test_bad_replica_condition

Sample Run

https://friendpaste.com/5zt2GTWpuYMoA23wf5lTRK

Change-Id: Iafb7aea69d666986c332143de0be6cbd790c82ca
Reviewed-on: http://review.couchbase.org/41179
Reviewed-by: Thuan Nguyen <soccon@gmail.com>
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Parag Agarwal <agarwal.parag@gmail.com>
Tested-by: Parag Agarwal <agarwal.parag@gmail.com>