MB-12251: Improve spatial tests
This commit makes the spatial tests easier to debug.

Every test run will return the exact same results. The documents
contents is always the same, even when random generated, it's
generated the same way.

The `num-docs` parameter is now mandatory (no default value is set).
It needs to be set in the test configuration.

Change-Id: I88e82f24bdf02c256403ede1b9ecdb17ae052a2b
Reviewed-on: http://review.couchbase.org/42600
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Ketaki Gangal <ketakigangal@gmail.com>