Reset the rebalance retry progess counter when we retrieve progress
We currently don't reset the rebalance retry progress counter and
this can lead to false test failures. Since this counter is
incremented every time the rebalance progress doesn't change it is
possible that the counter is incremented a few times, then progress
resumes then stops and is incremented one more time causing the
test to immediately fail. The retry counter should be reset to ensure
that the counter is incremented N consecutive times before the test
is marked failed.

Change-Id: Ib538bba937e3b4a8396ba660a40f79b9bb404be1
Reviewed-on: http://review.couchbase.org/44716
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Tommie McAfee <tommie@couchbase.com>