CBQE-2542 :: Add support for configuring single node cluster with data services
- We can provision 1 node cluster with data services
- These services can be defined via .ini file or services_init parameter
- Implemented for all tests inheriting from basetestcase.py
- works for dev box and remote VMs
- We cannot remove services from a node,
we can only remove a node from cluster, reinstall the server on it
- Once provisioned any attempt to provision the services will not succeed
but we will not consider it failure as in sherlock as we cannot remove the services.
The test cases that run after first test case, will face this issue and we avoid failures with this assumption
- Services = None is considered no provisioning the system will just pass.
This is to make sure we are backward compatible.
Change-Id: Ib14538d7f106575910206c75c6b13c01380b2238
Reviewed-on: http://review.couchbase.org/44995
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Tommie McAfee <tommie@couchbase.com>