MB-13665: Fix viewmerge test
For `updater_after` requests there's an internal stats cache within the view
engine. The current timeout is 0.3 seconds. This means that you might see
outdated update sequence numbers for up to 0.3 seconds. Hence add a sleep so
we can be sure the view engine retrieves the most current update sequence.

Change-Id: Icae1659d5fb3319e6305ad63f61997fa8a597675
Reviewed-on: http://review.couchbase.org/47467
Reviewed-by: Harsha H S <hhs.couchbase@gmail.com>
Tested-by: Volker Mische <volker.mische@gmail.com>