CBQE-2760 Compaction test fail to detect the bucket_compaction task
Compaction test asynchronously generate mutations in batches and check for fragmentation.
Once the desired fragmentation level is reached and the batch of mutations is
complete we check for the bucket_compaction task. However if the mutation batch
runs too long after the fragmentation level is reached then the check for the
bucket_compaction task is delayed and the bucket_compaction task may have completed
in this window. 

The fix is to decrease the granularity of the mutation batches by a factor of 10 so
that we are effectively polling the fragmentation level more frequently.

Change-Id: I8bdbc0cac6af1a61e1a3f555e91c410913a32648
Reviewed-on: http://review.couchbase.org/49686
Reviewed-by: Eric Cooper <ericcouchbase@gmail.com>
Tested-by: Eric Cooper <ericcouchbase@gmail.com>