CBQE-0: Add support for setting index ram quota
Now we can set index ram quota for GSI using test case
This uses a rest api call. We have added a parameter for this
purpose. index_quota_percent. This will be very
useful when defining DGM scenarios. Note that index ram quota
and memory quote don't add up the sum of the total ram available.

By default, if the parameter is not set, we will use the following
80% of 2/3 of memory available.

Please do not use this for any other release except >= 4.0.0.
This has been programmed to work in sherlock only.

Change-Id: I872a12e259aa5beccd408d7794d3e537c694c829
Reviewed-on: http://review.couchbase.org/50085
Reviewed-by: Ashvinder Singh <ashvinder@couchbase.com>
Tested-by: Ashvinder Singh <ashvinder@couchbase.com>