CBQE-2313: Increase the number of docs per vBucket
ns_server doesn't take the total number of documents into account when
determining whether to use a subset of the data for a query on a
development view or not, but only the documents on the local node. This
seem to include also replicas.

When the stat for the number of documents is requested, it might not
yet contain the full number. This means that when this test was running
we were sometimes getting below the threshold. Increasing the number of
documents be vBucket significantly should solve this issue.

Change-Id: Ib1eae8f5f042fb8fe3f8526a5ebc2690dff685f5
Reviewed-on: http://review.couchbase.org/59194
Reviewed-by: Harsha H S <hhs.couchbase@gmail.com>
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Abhishek Singh <abhishek@couchbase.com>
Reviewed-by: Ritam Sharma <ritamcouchbase@gmail.com>