Make view-engine test less fragile
The `test_stale_update_after_alternated_docs` test was using a sleep
to wait until the updater is finished. That's the wrong approach as
it may take longer than the sleep. Increasing the sleep also isn't
a proper solution as it would make the test run longer than needed.

The proper solution is to poll the result in a certain interval until
the expected result is returned. If not, the test fails after too many
retries.

Change-Id: Iee4f57a5e8fff142e5697cbc1b35f5f0ca6fc892
Reviewed-on: http://review.couchbase.org/60699
Reviewed-by: Abhishek Singh <abhishek@couchbase.com>
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Arunkumar Senthilnathan <arun.couchbase@yahoo.com>