Reconnect if disconnected during check for vbucket state
A client may be disconnected at any given time (it could
for instance try to connect to a bucket before it is
enabled and hence get the "no bucket" error and get
disconnected. This is not a fatal error in this function
so we could simply just reconnect and try again.

Change-Id: I665c38df54cdff0b6d45a2002d1390b043ca6afc
Reviewed-on: http://review.couchbase.org/64448
Reviewed-by: Dave Finlay <dave.finlay@couchbase.com>
Tested-by: Dave Finlay <dave.finlay@couchbase.com>