MB-22707: Capture the logs when using cluster_run
Some tests are run against a cluster set-up using cluster_run.  If
there is a test failure we would like to capture the logs so they can
be analysed off-line.

This patch introduces a new parameter get_logs_cluster_run which
if set to true, on a test failure will create a zip file of the
ns_server/logs directory and place the file in the testrunner log
directory.

Change-Id: Ifc07240b63562482dda0ff29ef7476cf044d7dea
Reviewed-on: http://review.couchbase.org/73574
Reviewed-by: Tommie McAfee <tommie@couchbase.com>
Tested-by: Daniel Owen <owend@couchbase.com>