CBQE-3984: reduce HTP memory usage
Merging of kv-store could lead to up to
4x memory usage while child-processes are
copying their kv-stores into the kv-store
of the main process.
This is problematic for tests loading +100k items.

The solution here is to not give each child process
it's own copy of the entire kv_store,
but allow the process to create a tmp_store
which get's merged later so that there is no overhead

Change-Id: I3a28fc1511d2927ebb9cdaa8ad2b60daa13e2549
Reviewed-on: http://review.couchbase.org/75086
Tested-by: Tommie McAfee <tommie@couchbase.com>
Reviewed-by: Eric Cooper <ericcouchbase@gmail.com>