always auth as an Administrator when doing bucket operations
this is a temporary measure to make tests pass after buckets security
will be tightened.

the real thing will be: when creating bucket, create bunch of users
with permissions for certain operations and use these users depending
of which operation is being performed. I'll get back to it later on

Change-Id: Ib62de122ac5f9e0147ccf215c67e66a2cadf4a70
Reviewed-on: http://review.couchbase.org/75736
Reviewed-by: Arunkumar Senthilnathan <arun.couchbase@yahoo.com>
Tested-by: Artem Stemkovski <artem@couchbase.com>