CBQE-5675: Fix docker dispatcher to call docker executor
Fixed the issue with dispatcher not dispatching jobs
based on dashboard descriptor params. The issue was that some job
dispatcher uses special dashboard parameters that are added to sub-component later on. Fixed the issue of finding if job was
run successfully before based on the dispatcher-parameters rather
than the sub-component given to the dispatcher.
Also fixed the issue of cleaning up the docker machines in case the
job isn't dispatched.
Also made the default option for fresh_run as true. Now passing the param will store false instead.

Change-Id: I54377133c8f47b1016bec85fe79d3f9b15635b71
Reviewed-on: http://review.couchbase.org/c/testrunner/+/126908
Reviewed-by: Ritesh Agarwal <ritesh.agarwal@couchbase.com>
Tested-by: Bharath G P <bharath.gp@couchbase.com>