Conditionally run cbbackupmgr with the required cloud arguments
By default the 'objstore_provider' will not be defined (None) so we
should see no change in behavior, however, once a provider is defined we
should see that cbbackupmgr is run with the required cloud arguments.

NOTE: The cloud arguments are all added when an 'objstore_provider' is
present. Extra conditions will need to be added when the time comes to
perform negative testing on the CLI parameters.

Change-Id: I47878cce36030dd3e9abf4a62a4f3b97b73acf11
Reviewed-on: http://review.couchbase.org/c/testrunner/+/130777
Reviewed-by: Patrick Varley <patrick@couchbase.com>
Reviewed-by: Carlos Gonzalez <carlos.gonzalez@couchbase.com>
Reviewed-by: Thuan Nguyen <soccon@gmail.com>
Tested-by: Thuan Nguyen <soccon@gmail.com>