[2/n] Conditionally run cbbackupmgr with the required cloud arguments
By default the 'objstore_provider' will not be defined (None) so we
should see no change in behavior, however, once a provider is defined we
should see that cbbackupmgr is run with the required cloud arguments.

NOTE: The cloud arguments are all added when an 'objstore_provider' is
present. Extra conditions will need to be added when the time comes to
perform negative testing on the CLI parameters.

Change-Id: Ib4712adb3a370cd35672c6f2baeb5027c9235f94
Reviewed-on: http://review.couchbase.org/c/testrunner/+/131073
Tested-by: Thuan Nguyen <soccon@gmail.com>
Reviewed-by: Thuan Nguyen <soccon@gmail.com>