[38/n] Support supplying objstore ssl arguments to cbbackupmgr
The 'cacert' and 'no_ssl_verify' values will now be read from the given
ini file and provided to boto3 and cbbackupmgr. With the correct
localstack setup (using the 'USE_SSL' environment variable) its now
possible to test '--obj-cacert' and '--obj-no-ssl-verify'.

NOTE: Given that testrunner expects localstack to be setup and provided
no test definitions will be provided to enable this testing since it
depends entirely on the localstack setup.

Change-Id: Ifcf6cbfc7d8e616469cd034c25b770bd71fe4ebb
Reviewed-on: http://review.couchbase.org/c/testrunner/+/131621
Reviewed-by: Carlos Gonzalez <carlos.gonzalez@couchbase.com>
Reviewed-by: Thuan Nguyen <soccon@gmail.com>
Tested-by: Thuan Nguyen <soccon@gmail.com>