CBQE-5955: test_backup_merge false positive fix.
The test 'test_backup_merge' was passing despite an unsuccessful merge.

This commit introduces 3 changes:

1. The test “test_backup_merge” was resulting in a false positive due to the
number of backups listed by the “cbbackupmgr info” command being counted
incorrectly. The count was only incremented if the backup exists in the list of
expected backups. If the backup doesn’t exist (an error condition), the counter
is not incremented. The counter is now incremented in both circumstances.

2. Panic detection in backup logs for backup_merge if the check_for_panic flag
is set.

3. Added additional functionality for checking for words in the backup logs and
the showing lines before and after (e.g. to capture stack traces).

Change-Id: I49e1a33f585f3ea5fdf9830c13035dc49d45da76
Reviewed-on: http://review.couchbase.org/c/testrunner/+/134087
Tested-by: Asad Zaidi <asad.zaidi@couchbase.com>
Reviewed-by: James Lee <james.lee@couchbase.com>
Reviewed-by: Arunkumar Senthilnathan <arun.couchbase@yahoo.com>