CBQE-6037: Amended backup_service_client to return http error codes
To make the test code easier to read, I amended the
backup_service_client to return http error codes as opposed to throwing
an ApiException on failure. This means we can write test cases which
simply check the http error code as opposed to checking if an exception
is thrown / not thrown.

Change-Id: Ia4ee63d407648af33f22c886812b89507a40f561
Reviewed-on: http://review.couchbase.org/c/testrunner/+/135712
Reviewed-by: Arunkumar Senthilnathan <arun.couchbase@yahoo.com>
Reviewed-by: Carlos Gonzalez <carlos.gonzalez@couchbase.com>
Tested-by: Asad Zaidi <asad.zaidi@couchbase.com>