CBQE-6107: Initialise bkrs_client as NoneType instead of as empty list.
The bkrs_client attribute is initialised as an empty list.

However the condition 'self.input.bkrs_client is not None', which is
present in various lines in testrunner, assumes self.input.bkrs_client
is initialised as None as opposed to an empty list.

This results in the condition 'self.input.bkrs_client is not None'
passing regardless of whether the bkrs_client is present in the ini file
or not.

Consequently testrunner passes an empty list to functions which expects
a TestInputServer

See CBQE-6107 for more details.

Change-Id: I5daa3c5738fb42f585e52dd74e120ce15e203406
Reviewed-on: http://review.couchbase.org/c/testrunner/+/137028
Reviewed-by: Thuan Nguyen <soccon@gmail.com>
Tested-by: Asad Zaidi <asad.zaidi@couchbase.com>