CBQE-6037: Shared directory and rebalance infrastructure amendments
A rebalance will only be triggered on the first backup service test and
an eject rebalance will only be triggered on the final test. This avoids
having to rebalance between test runs speeding things up. The tear down
process in between simply deletes the buckets and deletes any files used
by cbbackupmgr e.g. the archive directory. This change is also a
workaround to MB-41898.

Running an 'exportfs -ua' unexports all directories which prevents an
issue in which the client could not mount the shared directory due to a
stale file handle.

Added code to make sure the shared directory and client directories are
empty by deleting them and recreating them.

Amended 'test_importing_sensitive_repository' to work with shared
directories by modifying the permissions of the shared directory as
opposed to the client sided mounted directory.

Fixed a bug where the arguments to chmod were being passed the wrong
around causing the call to the import repository endpoint to return a
200 as opposed to a 400 in test_importing_read_only_repository.

Added a newline between libraries.

Change-Id: I50d5698459f9b5cd6175a92827306159afdde647
Reviewed-on: http://review.couchbase.org/c/testrunner/+/137821
Tested-by: Arunkumar Senthilnathan <arun.couchbase@yahoo.com>
Reviewed-by: Arunkumar Senthilnathan <arun.couchbase@yahoo.com>
Reviewed-by: Carlos Gonzalez <carlos.gonzalez@couchbase.com>