CBQE-6037: Amend bug in backup_cluster
The backup_cluster method appends a backup to a list of backups, it grabs
them from the filesystem by running 'ls' sorting in reverse order by
date modified and grabing the tail, however meta-data files may end up
in the directory listing. As we only require the backups can filter out
anything that's not a backup.

Change-Id: Ief838a46f6a1972bc7448ff507e697aebc8e0b9d
Reviewed-on: http://review.couchbase.org/c/testrunner/+/138242
Reviewed-by: Carlos Gonzalez <carlos.gonzalez@couchbase.com>
Reviewed-by: Arunkumar Senthilnathan <arun.couchbase@yahoo.com>
Tested-by: Arunkumar Senthilnathan <arun.couchbase@yahoo.com>