CBQE-6167: Avoid sharing and fetching info from staging directory
It is no longer possible to share staging directories between
cbbackupmgr and cbriftdump.

Consequently, the strategy to deal with this is to avoid fetching
backups and bucket-uuids from the staging directory and to fetch them
from the object store directly. This allows us to delete the staging
directory before and after cbriftdump is used.

Change-Id: I48936c58d79d36f208c2fa3fc065725f62359f64
Reviewed-on: http://review.couchbase.org/c/testrunner/+/138640
Reviewed-by: James Lee <james.lee@couchbase.com>
Tested-by: Asad Zaidi <asad.zaidi@couchbase.com>