CBQE-6167: Avoid sharing staging directories
This is a subset of the original due to Python3/Python2
incompatabilities, the compatability being that you cannot parse rfc3339
easily in Python2 without importing an external library.

This patch takes the following subset of the original patch:

It is no longer possible to share staging directories between
cbbackupmgr and cbriftdump.

Hence, we can use a temporary staging directory for cbriftdump to
prevent a clash with the staging directory used by cbbackupmgr.

Change-Id: I48936c58d79d36f208c2fa3fc065725f62359f64
Reviewed-on: http://review.couchbase.org/c/testrunner/+/138640
Reviewed-by: James Lee <james.lee@couchbase.com>
Tested-by: Asad Zaidi <asad.zaidi@couchbase.com>
Reviewed-on: http://review.couchbase.org/c/testrunner/+/140979
Reviewed-by: Thuan Nguyen <soccon@gmail.com>