CBQE-6037: (1/2) Infrastructure changes for schedule testing
A Repository can now determine which task is scheduled next and what
time it is scheduled.

Given a time, a Plan can determine the next task is scheduled. This is
coupled with logic to only to update specific tasks. This allows us to
build a schedule tester which follows similar logic to the backup
service.

Change-Id: I93ddc85f40e718c75cd4903c770945968f8e80d4
Reviewed-on: http://review.couchbase.org/c/testrunner/+/139612
Reviewed-by: Thuan Nguyen <soccon@gmail.com>
Reviewed-by: Carlos Gonzalez <carlos.gonzalez@couchbase.com>
Tested-by: Asad Zaidi <asad.zaidi@couchbase.com>