CBQE-6037: Fetch uid/gid dynamically and sort servers by free memory
Accesses to the NFS are mapped to the couchbase user and group. These
are hard-coded. In the situation where another user and group have
already taken these numbers, the NFS share will not work as intended. To
avoid running into this situation, we can fetch the uid and gid
dynamically.

In addition, sort the servers by available memory in ascending order so
the server with the least amount of available memory is at the start.

This is required because trying to add a node with less memory to a node
with more memory fails.

Change-Id: I57bc13d9b274750262da88e362fcbcf4831f91f9
Reviewed-on: http://review.couchbase.org/c/testrunner/+/140815
Reviewed-by: Thuan Nguyen <soccon@gmail.com>
Tested-by: Asad Zaidi <asad.zaidi@couchbase.com>