CBQE-6037: Move the sorting of servers by memory to the preamble
The servers are sorted in ascending order by memory to prevent the
situation in which a node with less memory is added to a node with more
memory. As the tearDown is called during the setUp, we can move the
sorting logic to the preamble as it's called at the start of both the
setUp and the tearDown.

Add a param to trigger the time reset in between tests as opposed to
performing it every test as performing time changes can break the stat
endpoint.

Change-Id: I1cab4e33e9265d819c4c5d3dc5f49b4624bbb1a5
Reviewed-on: http://review.couchbase.org/c/testrunner/+/140878
Reviewed-by: Thuan Nguyen <soccon@gmail.com>
Tested-by: Asad Zaidi <asad.zaidi@couchbase.com>