CBQE-6136: Amend history tests
As backups made in chesire-cat are always collection aware. Previously
this was not the case and when trying to examine a collection unaware
backup using a collection string, you would receive a warning. As it's
no longer possible to obtain this warning, I have amended this test to
check that a backup which does not explicitly contain a collection can
be examined using a bucket only string and also with a bucket suffixed
with '_default._default'.

Amend a test in the conf file where test name was spelled incorrectly.

Change-Id: I49910c75addd262bb0e49b4958faac9a565f06df
Reviewed-on: http://review.couchbase.org/c/testrunner/+/141061
Reviewed-by: Thuan Nguyen <soccon@gmail.com>
Tested-by: Asad Zaidi <asad.zaidi@couchbase.com>