CBQE-6037: Configure the master node correctly
Background:

Initially, the master node is selected as the first node from the list
`servers`.

To alleviate the issue of adding a node with lower memory to a node with
higher memory, the `cluster` list is sorted by available memory.

A new master node is elected as the first element in the cluster list
following the sort.

The objects present in the `server` list and the `cluster` list are
different instances of a TestInputServer with the same attribute values.

As there is no __eq__ method defined (e.g. to compare TestInputServers
by attribute values), comparisons throughout testrunner will fail if you
select a master server from the cluster list as opposed to the server
list.

The Fix:

To alleviate this issue without amending testrunner code, we can just
elect a new master node from the list of `servers` as opposed to the
`cluster` list.

Change-Id: Ic862a9516d4979f1548ae827f822d5d37dfdf3d5
Reviewed-on: http://review.couchbase.org/c/testrunner/+/141235
Tested-by: Arunkumar Senthilnathan <arun.couchbase@yahoo.com>
Reviewed-by: Arunkumar Senthilnathan <arun.couchbase@yahoo.com>