CBQE-6239: Add a space between keys and values in regular expressions.
The fix:

Add a space between keys and values in regular expressions in'conf/entbackuprestore/entbackuprestore-regex-ssl.conf'

The why:

The primary reason is that cbbackupmgr requires the space in between the key and values as the unmarshalled documents it observes has spaces in between keys and values.

See CBQE-6239 for more details.

Change-Id: Ic7419c811b9140dcdbb361ed946adb48de512e26
Reviewed-on: http://review.couchbase.org/c/testrunner/+/141396
Reviewed-by: Thuan Nguyen <soccon@gmail.com>
Tested-by: Asad Zaidi <asad.zaidi@couchbase.com>