CBQE-6240: A flag to place a bucket in the disk greater than memory state
The `dgm_run` input param can be set which places a bucket in the dgm
state.

My understanding is that each bucket has a specific quota, when
documents are inserted such that it exceeds the allocated quota, at
which point the residency ratio decreases below 100% as you cannot store
all the documents in disk in memory.

Change-Id: I96c6f461b492ffe1fe90c0fd7f24e272faf8044b
Reviewed-on: http://review.couchbase.org/c/testrunner/+/141441
Tested-by: Arunkumar Senthilnathan <arun.couchbase@yahoo.com>
Reviewed-by: Arunkumar Senthilnathan <arun.couchbase@yahoo.com>