CBQE-6243: Test the backup service lots of scopes/collections
Test the backup service can restore lots of collections.

It's currently challenging to test with a 1000 collections as testing
on single core vms makes the disk write queue very slow.

Creating a large number collections and attempting to flush the bucket
via the doFlush endpoint returns an internal error. This in fact
documented behaviour and happens when the flush takes more than 30
seconds to complete. Dropping a bucket appears to be faster so to avoid
this situation, I will drop the bucket as opposed to flushing the
bucket.

There's a delay between a restore succeeding and all documents being
accessible on memcached, to avoid this situation I will add a large
sleep after a restore.

In the case of a single document being inserted into each collection,
it's possible to have all documents end up in a single vbucket. To avoid
this, we can prepend the collection id to each document key.

Change-Id: Icefbe179d4f6f307fb696bfab6fa3a06ee30dc48
Reviewed-on: http://review.couchbase.org/c/testrunner/+/141605
Tested-by: Arunkumar Senthilnathan <arun.couchbase@yahoo.com>
Reviewed-by: Arunkumar Senthilnathan <arun.couchbase@yahoo.com>